+++ src/main/java/com/owlmaddie/mixin/MixinMobEntity.java
@@ -1,24 +1,21 @@
 package com.owlmaddie.mixin;
 
 import com.owlmaddie.chat.ChatDataManager;
-import com.owlmaddie.commands.ConfigurationHandler;
 import com.owlmaddie.network.ServerPackets;
 import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.entity.passive.TameableEntity;
+import net.minecraft.entity.passive.VillagerEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.registry.Registries;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
-import net.minecraft.util.Identifier;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import java.util.List;
 
 /**
  * The {@code MixinMobEntity} mixin class exposes the goalSelector field from the MobEntity class.
@@ -30,6 +27,11 @@ public class MixinMobEntity {
     private void onItemGiven(PlayerEntity player, Hand hand, CallbackInfoReturnable<ActionResult> cir) {
         ItemStack itemStack = player.getStackInHand(hand);
         MobEntity thisEntity = (MobEntity) (Object) this;
+
+        // Don't interact with Villagers (avoid issues with trade UI) OR Tameable (i.e. sit / no-sit)
+        if (thisEntity instanceof VillagerEntity || thisEntity instanceof TameableEntity) {
+            return;
+        }
 
         // Determine if the item is a bucket
         // We don't want to interact on buckets
@@ -48,26 +50,35 @@ public class MixinMobEntity {
             return;
         }
 
+        // Get chat data for entity
+        ChatDataManager chatDataManager = ChatDataManager.getServerInstance();
+        ChatDataManager.EntityChatData chatData = chatDataManager.getOrCreateChatData(thisEntity.getUuidAsString());
+
         // Check if the player successfully interacts with an item
-        if (!itemStack.isEmpty() && player instanceof ServerPlayerEntity) {
-            ServerPlayerEntity serverPlayer = (ServerPlayerEntity) player;
-            String itemName = itemStack.getItem().getName().getString();
-            int itemCount = itemStack.getCount();
+        if (player instanceof ServerPlayerEntity) {
+            // Player has item in hand
+            if (!itemStack.isEmpty()) {
+                ServerPlayerEntity serverPlayer = (ServerPlayerEntity) player;
+                String itemName = itemStack.getItem().getName().getString();
+                int itemCount = itemStack.getCount();
 
-            // Decide verb
-            String action_verb = " shows ";
-            if (cir.getReturnValue().isAccepted()) {
-                action_verb = " gives ";
-            }
+                // Decide verb
+                String action_verb = " shows ";
+                if (cir.getReturnValue().isAccepted()) {
+                    action_verb = " gives ";
+                }
 
-            // Prepare a message about the interaction
-            String giveItemMessage = "<" + serverPlayer.getName().getString() +
-                    action_verb + "you " + itemCount + " " + itemName + ">";
+                // Prepare a message about the interaction
+                String giveItemMessage = "<" + serverPlayer.getName().getString() +
+                        action_verb + "you " + itemCount + " " + itemName + ">";
 
-            ChatDataManager chatDataManager = ChatDataManager.getServerInstance();
-            ChatDataManager.EntityChatData chatData = chatDataManager.getOrCreateChatData(thisEntity.getUuidAsString());
-            if (!chatData.characterSheet.isEmpty() && chatData.auto_generated < chatDataManager.MAX_AUTOGENERATE_RESPONSES) {
-                ServerPackets.generate_chat("N/A", chatData, serverPlayer, thisEntity, giveItemMessage, true);
+                if (!chatData.characterSheet.isEmpty() && chatData.auto_generated < chatDataManager.MAX_AUTOGENERATE_RESPONSES) {
+                    ServerPackets.generate_chat("N/A", chatData, serverPlayer, thisEntity, giveItemMessage, true);
+                }
+
+            } else if (itemStack.isEmpty() && chatData.friendship == 3) {
+                // Player's hand is empty, Ride your best friend!
+                player.startRiding(thisEntity, true);
             }
         }
     }