+++ src/main/java/com/owlmaddie/network/ServerPackets.java
@@ -7,7 +7,6 @@ import com.owlmaddie.goals.EntityBehaviorManager;
 import com.owlmaddie.goals.GoalPriority;
 import com.owlmaddie.goals.TalkPlayerGoal;
 import com.owlmaddie.utils.Compression;
-import com.owlmaddie.utils.LivingEntityInterface;
 import com.owlmaddie.utils.Randomizer;
 import com.owlmaddie.utils.ServerEntityFinder;
 import io.netty.buffer.Unpooled;
@@ -42,7 +41,7 @@ import java.util.concurrent.TimeUnit;
 public class ServerPackets {
     public static final Logger LOGGER = LoggerFactory.getLogger("creaturechat");
     public static MinecraftServer serverInstance;
-    private static ChatDataSaverScheduler scheduler = null;
+    public static ChatDataSaverScheduler scheduler = null;
     public static final Identifier PACKET_C2S_GREETING = new Identifier("creaturechat", "packet_c2s_greeting");
     public static final Identifier PACKET_C2S_READ_NEXT = new Identifier("creaturechat", "packet_c2s_read_next");
     public static final Identifier PACKET_C2S_SET_STATUS = new Identifier("creaturechat", "packet_c2s_set_status");
@@ -217,17 +216,6 @@ public class ServerPackets {
                 scheduler.stopAutoSaveTask();
             }
         });
-
-        ServerEntityEvents.ENTITY_LOAD.register((entity, world) -> {
-            String entityUUID = entity.getUuidAsString();
-            if (ChatDataManager.getServerInstance().entityChatDataMap.containsKey(entityUUID)) {
-                int friendship = ChatDataManager.getServerInstance().entityChatDataMap.get(entityUUID).friendship;
-                if (friendship > 0) {
-                    LOGGER.info("Entity loaded (" + entityUUID + "), setting friendship to " + friendship);
-                    ((LivingEntityInterface)entity).setCanTargetPlayers(false);
-                }
-            }
-        });
         ServerEntityEvents.ENTITY_UNLOAD.register((entity, world) -> {
             String entityUUID = entity.getUuidAsString();
             if (entity.getRemovalReason() == Entity.RemovalReason.KILLED && ChatDataManager.getServerInstance().entityChatDataMap.containsKey(entityUUID)) {
@@ -276,20 +264,27 @@ public class ServerPackets {
 
         // Grab random adjective
         String randomAdjective = Randomizer.getRandomMessage(Randomizer.RandomType.ADJECTIVE);
-        String randomFrequency = Randomizer.getRandomMessage(Randomizer.RandomType.FREQUENCY);
+        String randomClass = Randomizer.getRandomMessage(Randomizer.RandomType.CLASS);
+        String randomAlignment = Randomizer.getRandomMessage(Randomizer.RandomType.ALIGNMENT);
+        String randomSpeakingStyle = Randomizer.getRandomMessage(Randomizer.RandomType.SPEAKING_STYLE);
 
+        // Generate random name parameters
+        String randomLetter = Randomizer.RandomLetter();
+        int randomSyllables = Randomizer.RandomNumber(5) + 1;
+
+        // Build the message
         StringBuilder userMessageBuilder = new StringBuilder();
-        userMessageBuilder.append("Please generate a " + randomFrequency + " " + randomAdjective);
-        userMessageBuilder.append(" character ");
+        userMessageBuilder.append("Please generate a ").append(randomAdjective).append(" character. ");
+        userMessageBuilder.append("This character is a ").append(randomClass).append(" class, who is ").append(randomAlignment).append(". ");
         if (entity.getCustomName() != null && !entity.getCustomName().getString().equals("N/A")) {
-            userMessageBuilder.append("named '").append(entity.getCustomName().getString()).append("' ");
+            userMessageBuilder.append("Their name is '").append(entity.getCustomName().getString()).append("'. ");
         } else {
-            userMessageBuilder.append("whose name starts with the letter '").append(Randomizer.RandomLetter()).append("' ");
-            userMessageBuilder.append("and uses ").append(Randomizer.RandomNumber(4) + 1).append(" syllables ");
+            userMessageBuilder.append("Their name starts with the letter '").append(randomLetter)
+                    .append("' and is ").append(randomSyllables).append(" syllables long. ");
         }
-        userMessageBuilder.append("and speaks in '" + userLanguage + "'" );
-        LOGGER.info(userMessageBuilder.toString());
+        userMessageBuilder.append("They speak in '").append(userLanguage).append("' with a ").append(randomSpeakingStyle).append(" style.");
 
+        LOGGER.info(userMessageBuilder.toString());
         chatData.generateMessage(userLanguage, player, "system-character", userMessageBuilder.toString(), false);
     }
 
@@ -315,6 +310,7 @@ public class ServerPackets {
                     LOGGER.debug("Setting entity name to " + characterName + " for " + chatData.entityId);
                     entity.setCustomName(Text.literal(characterName));
                     entity.setCustomNameVisible(true);
+                    entity.setPersistent();
                 }
 
                 PacketByteBuf buffer = new PacketByteBuf(Unpooled.buffer());